Move checkbox out of form/ folder, tighten styles
Moved the checkbox component to the root components/ folder but kept
form/checkbox for backwards compatibility. Also tightened up the styles.
